move types into separate folder
