+++ src/utils/types/DayWeather.ts
